% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_countries.R
\name{appbot_countries}
\alias{appbot_countries}
\title{appbot_countries}
\usage{
appbot_countries(app_id)
}
\arguments{
\item{app_id}{The ID of the app you want to check}
}
\description{
Gives a list of countries you can access reviews / ratings from
}
