% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_ratings.R
\name{appbot_ratings_android}
\alias{appbot_ratings_android}
\title{appbot_ratings_android}
\usage{
appbot_ratings_android(app_id, method, start = NULL, end = NULL)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{method}{Either 'historical' or 'live'}

\item{start}{The start date to search in}

\item{end}{The end date to search in}
}
\description{
Get ratings for an Android app
}
