% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_ratings.R
\name{appbot_ratings_ios}
\alias{appbot_ratings_ios}
\title{appbot_ratings_ios}
\usage{
appbot_ratings_ios(app_id, method, output, country, start = NULL, end = NULL)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{method}{Either 'historical' or 'live'}

\item{output}{Either 'current_version' or 'all_time'}

\item{country}{The two letter country code that is used in iOS, like 'DE' or 'IT'}

\item{start}{The start date to search in}

\item{end}{The end date to search in}
}
\description{
Get ratings for an iOS app
}
