% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_review_summary.R
\name{appbot_review_summary}
\alias{appbot_review_summary}
\title{appbot_review_summary}
\usage{
appbot_review_summary(
  app_id,
  method,
  start = NULL,
  end = NULL,
  country = NULL,
  sentiments = NULL,
  dlangs = NULL,
  version = NULL,
  topic = NULL,
  pad_empty = NULL
)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{method}{The review summary method you want to check out - possibilities are 'date', 'country', and 'rating'}

\item{start}{The start date you want to call in YYYY-MM-DD format}

\item{end}{The end date you want to call in YYYY-MM-DD format}

\item{country}{The ID of the country to filter by}

\item{sentiments}{Sentiments to filter by - could be one or more of 'positive', 'negative', 'neutral', and/or 'mixed'}

\item{dlangs}{One or more detected languages, using the language ID}

\item{version}{The version number to filter by}

\item{topic}{The topic, custom topic or tag ID to filter to}

\item{pad_empty}{Return items for empty days; should be either 'n' or 'y'}
}
\description{
Allows you to call different Appbot API endpoints related to review summaries
}
