% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_reviews.R
\name{appbot_reviews}
\alias{appbot_reviews}
\title{appbot_reviews}
\usage{
appbot_reviews(app_id, pages = 1, pivot_topics = FALSE, ...)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{pages}{How many pages of results you want to provide}

\item{pivot_topics}{An option to pivot the topics column (which is returned as a list) out to individual columns with a Boolean 1 or 0 to indicate if the review discusses a certain topic}

\item{...}{Other arguments include 'page', 'start', 'end', 'keyword', 'country', 'sentiments', 'dlangs', 'version', and 'topic'. For example if you wanted to get reviews from Germany, you would pass 'country=DE' as an argument}
}
\description{
Get reviews for the app of your choice
}
