% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_reviews_import.R
\name{appbot_reviews_import}
\alias{appbot_reviews_import}
\title{appbot_reviews_import}
\usage{
appbot_reviews_import(app_id, reviews_df)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{reviews_df}{A data frame of reviews that you want to import - must contain the following columns:
\itemize{
  \item id A unique identifier for each review
  \item store_id The app store ID - you can get this from \code{\link{appbot_reviews}}
  \item author The review author
  \item rating The review rating
  \item body The review text body
  \item subject The review subject
  \item published_at The review publishing date
  \item version The app version (iOS only)
  \item country_code The country code (see appbot_countries for codes)
}}
}
\description{
Allows you to call different Appbot API endpoints related to sentiments
}
