% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_topics_reviews.R
\name{appbot_topics_reviews}
\alias{appbot_topics_reviews}
\title{appbot_topics_reviews}
\usage{
appbot_topics_reviews(
  app_id,
  topic_id,
  pivot_topics = FALSE,
  start = NULL,
  end = NULL,
  country = NULL,
  sentiments = NULL,
  dlangs = NULL,
  version = NULL
)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{topic_id}{The topics ID you want to check out - needs to be a numeric value}

\item{pivot_topics}{An option to pivot the topics column (which is returned as a list) out to individual columns with a Boolean 1 or 0 to indicate if the review discusses a certain topic}

\item{start}{The start date you want to call in YYYY-MM-DD format}

\item{end}{The end date you want to call in YYYY-MM-DD format}

\item{country}{The ID of the country to filter by}

\item{sentiments}{Sentiments to filter by - could be one or more of 'positive', 'negative', 'neutral', and/or 'mixed'}

\item{dlangs}{One or more detected languages, using the language ID}

\item{version}{The version number to filter by}
}
\description{
Allows you to call the Appbot API to get reviews by topic
}
