% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_words.R
\name{appbot_words}
\alias{appbot_words}
\title{appbot_words}
\usage{
appbot_words(
  app_id,
  method,
  start = NULL,
  end = NULL,
  country = NULL,
  sentiments = NULL,
  dlangs = NULL,
  version = NULL,
  topic = NULL
)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{method}{The words method you want to check out - possibilities are 'popular', 'critical', 'interesting', 'new', 'trending_up', and 'trending_down'}

\item{start}{The start date you want to call in YYYY-MM-DD format}

\item{end}{The end date you want to call in YYYY-MM-DD format}

\item{country}{The ID of the country to filter by}

\item{sentiments}{Sentiments to filter by - could be one or more of 'positive', 'negative', 'neutral', and/or 'mixed'}

\item{dlangs}{One or more detected languages, using the language ID}

\item{version}{The version number to filter by}

\item{topic}{The topic, custom topic or tag ID to filter to}
}
\description{
Allows you to call different Appbot API endpoints related to keywords
}
