% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{start_end_sequencer}
\alias{start_end_sequencer}
\title{start_end_sequencer}
\usage{
start_end_sequencer(df, increment = 100)
}
\arguments{
\item{df}{The dataframe you want to sequence}

\item{increment}{How much you want your sequence to increment by. Defaults to 100}
}
\description{
Get a set of sequences for starting and ending a sequence
}
