% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appbot_reviews_import.R
\name{appbot_reviews_import}
\alias{appbot_reviews_import}
\title{appbot_reviews_import}
\usage{
appbot_reviews_import(
  app_id,
  store_id,
  author = NULL,
  rating = NULL,
  body = NULL,
  subject = NULL,
  version = NULL,
  country_code = NULL,
  published_at = NULL
)
}
\arguments{
\item{app_id}{The ID of the app you want to check}

\item{store_id}{The ID of the app store you are importing to}

\item{author}{The review author}

\item{rating}{The review rating}

\item{body}{The review text body}

\item{subject}{The review subject}

\item{version}{The app version (iOS only)}

\item{country_code}{The country code (see appbot_countries for codes)}

\item{published_at}{The review publishing date - if not specified, uses the current date time}
}
\description{
Allows you to call different Appbot API endpoints related to sentiments
}
